pyproject.toml
[tool.poetry]
name = "bim-streamlit"
version = "0.1.0"
description = "bim robot"
authors = ["amitca71 <amitca71@gmail.com>"]
license = "apache2"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.2.14"
langchain-openai = "^0.1.21"
langchain-community = "^0.2.12"
langchainhub = "^0.1.21"
neo4j = "^5.23.1"
retry = "^0.9.2"
segment-analytics-python = "^2.3.2"
streamlit = "^1.37.1"
streamlit-feedback = "^0.1.3"
langchain-google-genai = "^1.0.8"
networkx = "^3.3"
pyvis = "^0.3.2"
tiktoken = "^0.7.0"
pinecone-client = "5.0.1"
#langchain-pinecone = "^0.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
